+++ data/__init__.py
@@ -3,7 +3,7 @@
 from data.base_dataset import BaseDataset, DataExtractor
 from data.data_scaling import DataScaler, ZScale, LinScale, NoScale
 from data.data_splitting import DataSplitter
-from data.prepared_dataset_backup import PreparedDataset, CustomDataset, CustomClassificationDataset, CustomSampler
+from data.prepared_dataset import PreparedDataset, CustomDataset, CustomClassificationDataset, CustomSampler
 from data.raw_data_coversion import RawDataConverter
 
 __all__ = [
