+++ data/prepared_dataset.py
@@ -3,34 +3,189 @@
 PreparedDataset
 ---------------
 
-Drop-in optimized version of the KnowIt prepared dataset module.
+The ``PreparedDataset`` represents a ``BaseDataset`` that is preprocessed for model training.
+It inherits from ``BaseDataset``. Based on the provided path(s) to a dataset, it will populate the
+parent's variables.
 
-Key optimizations:
-- Avoid mutating self.batch_sampling_mode in get_dataloader()
-- Support DataLoader batch fetching via __getitems__()
-- Cache per-batch slice loads to avoid repeated I/O / pandas slicing
-- Preload NumPy arrays directly when preload=True
-- Better DataLoader defaults: pin_memory, persistent_workers, prefetch_factor
+-----------------
+Prediction points
+-----------------
+
+In order to define a ``PreparedDataset``, the input-output dynamics of the model to be trained
+must be defined rather precisely. The ``PreparedDataset`` is built on the idea of 'prediction points'.
+Each time step in the ``BaseDataset`` can be regarded a prediction point (under some assumptions).
+At every prediction point, a model is to predict a specific set of features-over-time from
+a specific other set of features-over-time. The specifics are defined as follows.
+
+    in_components : list
+        A subset of ``BaseDataset.components`` representing the components that will be used as input to the model.
+    out_components : list
+        A subset of ``BaseDataset.components`` representing the components that will be used as output to the model.
+    in_chunk : list
+        A list of two integers `[a, b]` for which `a <= b` defining the time steps (of `in_components`)
+        to be used for prediction at point `t` as `[t + a, t + b]`.
+    out_chunk : list
+        A list of two integers `[a, b]` for which `a <= b` defining the time steps (of `out_components`)
+        to be predicted at prediction point `t` as `[t + a, t + b]`.
+
+Note that this might seem cumbersome, but it allows us to easily define several different types of tasks.
+For example:
+
+- regression (predict heartrate from an 11 time-step window given three instantaneous biometrics)
+    - in_components = [biometric1, biometric2, biometric2]
+    - out_components = [heart rate]
+    - in_chunk = [-5, 5]
+    - out_chunk = [0, 0]
+- autoregressive univariate forecasting (predict a stock's value in 5 time steps given last 21 time steps)
+    - in_components = [stock,]
+    - out_components = [stock,]
+    - in_chunk = [-20, 0]
+    - out_chunk = [5, 5]
+- time series classification (detect a whale call given an audio recording of 5 time steps)
+    - in_components = [sound,]
+    - out_components = [positive_call,]
+    - in_chunk = [-2, 2]
+    - out_chunk = [0, 0]
+
+--------------------------
+Variable length processing
+--------------------------
+
+The ``PreparedDataset`` can also be defined in a "variable length" configuration.
+This means that instead of receiving and producing chunks of a specific number of time steps,
+the model will receive a variable number of time steps and produce an identical number of time steps as output.
+The input- and output components don't need to be the same. Not all architectures can facilitate this configuration.
+Currently, only ``LSTMv2``, ``TCN``, and ``CNN`` supports it.
+
+This configuration is selected by defining ``task=vl_regression``, ``batch_sampling_mode=variable_length``,
+selecting an appropriate architecture, and ensuring that the size of ``in_chunk`` and ``out_chunk`` are both 1.
+Classification tasks are not currently supported for this configuration.
+
+Examples of tasks that can be defined with this configuration include:
+
+- regression (predict heartrate from the change in three instantaneous biometrics over time)
+    - in_components = [biometric1, biometric2, biometric2]
+    - out_components = [heart rate]
+    - in_chunk = [0, 0]
+    - out_chunk = [0, 0]
+- autoregressive univariate forecasting (predict a stock's value in 5 time steps given its value in the past)
+    - in_components = [stock,]
+    - out_components = [stock,]
+    - in_chunk = [0, 0]
+    - out_chunk = [5, 5]
+
+
+--------------------
+Splitting & Limiting
+--------------------
+
+The first step in preparing the dataset is to split it into a train-, validation-,
+and evaluation set. This is done with the ``DataSplitter`` module.
+The ``split_method`` data keyword argument defines the way the dataset is split.
+More details can be found in ``DataSplitter``, but we summarize the options here:
+    - 'random': Ignore all distinction between instances and slices, and split on time steps randomly.
+    - 'chronological' (default): Ignore all distinction between instances and slices, and split on time steps chronologically.
+    - 'slice-random': Ignore all distinction between instances, and split on slices randomly.
+    - 'slice-chronological': Ignore all distinction between instances, and split on slices chronologically.
+    - 'instance-random': Split on instances randomly.
+    - 'instance-chronological': Split on instances chronologically.
+    - 'custom': User defined split.
+Note that the data is split ON the relevant level (instance, slice, or timesteps).
+I.e. If you split on instances and there are only 3 instances, then split_portions=(0.6, 0.2, 0.2)
+will be a wild approximation i.t.o actual time steps.
+
+Note that, if desired, the data is limited during splitting, and the data is limited by removing
+the excess data points from the end of the data block after shuffling or ordering according to time.
+Also note that if the data is limited too much for a given ``split_portion`` to have a single entry,
+an error will occur confirming it.
+
+Note that a 'custom' split can be constructed during the data importing.
+See the ``RawDataConverter`` module for more information.
+
+-------
+Scaling
+-------
+
+After the data is split, a scaler is fit to the train set data
+which will be applied to all data being extracted for model training or inference.
+This is done with the ``DataScaler`` module and the corresponding ``scaling_method`` and ``scaling_tag``,
+data keyword arguments. More details can be found there, but we summarize the options here:
+    - scaling_method='z-norm': Features are scaled by subtracting the mean and dividing by the standard deviation.
+    - scaling_method='zero-one': Features are scaled linearly to be in the range (0, 1).
+    - scaling_method=None: No scaling occurs.
+    - scaling_tag='in_only': Only the input features will be scaled.
+    - scaling_tag='full': The input and output features will be scaled.
+    - scaling_tag=None: No features will be scaled.
+Note that scaling of output components is not permitted if performing a classification task. ``scaling_tag``='full' will be
+automatically changed to ``scaling_tag``='in_only' for classification tasks.
+Also note that the scaling happens "online" as data is sampled from disk.
+
+-------
+Padding
+-------
+
+At some prediction points the ``in_chunk`` might exceed the corresponding
+slice range. In these cases we pad the input values. The output values are never padded.
+Prediction points that do not have valid output values for an ``out_chunk`` are not selected
+as appropriate prediction point during data splitting. The argument ``padding_method`` is the same
+as ``mode`` in the numpy.pad function (https://numpy.org/doc/stable/reference/generated/numpy.pad.html).
+See the ``DataSplitter`` module for more information.
+
+-------------
+CustomDataset
+-------------
+
+``PreparedDataset.get_dataloader()`` can be called to return a Pytorch dataloader that can be used for model training.
+This function compiles a dataloader with a ``CustomDataset`` class.
+This class inherits from ``torch.utils.data.Dataset``.
+It defines how the samples will look when enumerating over the dataloader.
+This includes sampling, padding, and normalizing.
+
+---------------------------
+CustomClassificationDataset
+---------------------------
+
+If the task is a classification task a ``CustomClassificationDataset`` will be used instead of ``CustomDataset``.
+``CustomClassificationDataset`` inherits from ``CustomDataset`` and adds some classification specific methods.
+
+-------------------------------------
+CustomVariableLengthRegressionDataset
+-------------------------------------
+
+If the task is a variable length regression task a ``CustomVariableLengthRegressionDataset`` will be used instead of ``CustomDataset``.
+``CustomVariableLengthRegressionDataset`` inherits from ``CustomDataset`` and adds some variable length regression specific methods.
+
+-------------
+CustomSampler
+-------------
+
+In addition to ``CustomDataset``, ``PreparedDataset`` also uses a custom batch sampler ``CustomSampler``
+when ``PreparedDataset.get_dataloader()`` is called to generate a dataloader.
+This class supports three different modes of temporal contiguity.
+    - 'independent': Time is contiguous within sequences (i.e. prediction points) but not enforced within or across batches.
+    - 'sliding-window': Time is contiguous within sequences and across batches, as far as possible.
+    - 'variable_length': Time is contiguous within sequences and across batches, as far as possible. Batches are also variable (but uniform) in length.
+See the ``CustomSampler`` module for details.
+
 """
 
 from __future__ import annotations
 __copyright__ = 'Copyright (c) 2025 North-West University (NWU), South Africa.'
 __licence__ = 'Apache 2.0; see LICENSE file for details.'
-__author__ = 'tiantheunissen@gmail.com, optimized by ChatGPT'
-__description__ = ('Contains the PreparedDataset, CustomSampler, CustomDataset, and '
-                   'CustomClassificationDataset, and CustomVariableLengthRegressionDataset '
-                   'class for KnowIt.')
+__author__ = 'tiantheunissen@gmail.com'
+__description__ = ('Contains the PreparedDataset, CustomSampler, CustomDataset, and CustomClassificationDataset, '
+                   'and CustomVariableLengthRegressionDataset class for Knowit.')
 
-from numpy import (
-    array, random, unique, pad, isnan, arange, expand_dims, concatenate,
-    diff, where, split, ndarray, isscalar, asarray
-)
+# external imports
+from numpy import (array, random, unique, pad, isnan, arange, expand_dims, concatenate,
+                   diff, where, split, ndarray, isscalar, asarray)
 from numpy.random import Generator
 from pandas import isna, DataFrame
 from torch.utils.data import Dataset, DataLoader, Sampler
 from torch import from_numpy, is_tensor, Tensor, unsqueeze
 from torch import zeros as zeros_tensor
 
+# internal imports
 from data.base_dataset import BaseDataset
 from data.data_splitting import DataSplitter
 from data.data_scaling import DataScaler
@@ -40,6 +195,100 @@ logger = get_logger()
 
 
 class PreparedDataset(BaseDataset):
+    """This is the PreparedDataset which represents a dataset that is preprocessed for model training.
+    It contains all the variables in BaseDataset in addition to metadata regarding
+    data splitting, scaling, and sampling.
+
+    Parameters
+    ----------
+    meta_path : str
+        The path to the desired dataset metadata. The path should point to a pickle file.
+    package_path : str
+        The path to the desired dataset package. The path should point to a directory containing the
+        partitioned parquet.
+    in_components : list, shape=[num_in_components,]
+        A subset of the 'components' variable of the BaseDataset to be used as input to the model.
+    out_components : list, shape=[num_in_components,]
+        A subset of the 'components' variable of the BaseDataset to be used as output to the model.
+    in_chunk : list, shape=[2,]
+        A list of two integers [a, b] for which a <= b,
+        defining the time steps of in_components for each prediction point.
+    out_chunk : list, shape=[2,]
+        A list of two integers [a, b] for which a <= b,
+        defining the time steps of out_components for each prediction point.
+    split_portions : tuple, shape=[3,]
+        The approximate portions of the (train, valid, eval) splits. Needs to add up to 1.
+    seed : int
+        The seed for reproducibility.
+    batch_size : int
+        The mini-batch size for training.
+    split_method : str
+        The method of splitting data. Options are 'random', 'chronological',
+        'instance-random', 'instance-chronological', 'slice-random', 'slice-chronological', or 'custom'.
+        See heading for description.
+    scaling_method : str | None
+        The method for scaling data features. Options are 'z-norm', 'zero-one', or None.
+        See heading for description.
+    scaling_tag : str | None
+        The mode to scale the data. Options are 'in_only', 'full', or None.
+        See heading for description.
+    shuffle_train : bool
+        Whether the training set is shuffled after every epoch.
+    limit : int | None
+        The number of elements (instances/slices/time) steps to limit the data to.
+        See heading for description.
+    padding_method : str
+        The method to pad model inputs with.
+        Options can be found at (https://numpy.org/doc/stable/reference/generated/numpy.pad.html).
+        See heading for description.
+    min_slice : int | None
+        The minimum slice size to consider during data splitting/selection.
+        If None, no slice selection is performed.
+    batch_sampling_mode : str | None
+        The sampling mode for generating batches in the CustomSampler class.
+        Either 'independent', 'sliding-window', or 'variable_length', as described in the CustomSampler module.
+    slide_stride : int
+        The stride used for the sliding-window approach, if selected.
+    variable_sequence_length_limit : int | None
+        If non-None value given and 'batch_sampling_mode' is 'variable_length',
+        the sequence lengths within a batch will be limited to the given value.
+
+    Attributes
+    ----------
+    x_map : array, shape=[n_in_components,]
+        An array that contains the indices of BaseDataset.components that correspond to input components.
+    y_map : array, shape=[n_out_components,]
+        An array that contains the indices of BaseDataset.components that correspond to output components.
+    train_set_size : int
+        The number of prediction points in the training set.
+    valid_set_size : int
+        The number of prediction points in the validation set.
+    eval_set_size : int
+        The number of prediction points in the evaluation set.
+    selection : dict[str, array]
+        A dictionary containing the selection matrices corresponding to each data split.
+    x_scaler : object
+        The scaler fitted to the training set input features.
+    y_scaler : object
+        The scaler fitted to the training set output features.
+    in_shape : list, shape=[n_input_time_delays, n_in_components]
+        A list that represents the shape of the inputs to the model.
+    out_shape : list, shape=[n_output_time_delays, n_output_components]
+        A list that represents the shape of the outputs to the model. Is modified for classification tasks.
+    class_set : dict
+        A dictionary that maps each class name to an integer class ID.
+        Only created if task='classification'.
+    class_counts : dict
+        A dictionary that maps each class ID to its size.
+        Only created if task='classification'.
+    custom_splits: dict
+        A dictionary defining the custom selection matrices.
+
+    Notes
+    -----
+        - All the listed parameters are also stored as attributes.
+        - The `_prepare` method is automatically called at initialization to initiate data preparation.
+    """
     # to be provided
     in_components = None
     out_components = None
@@ -70,15 +319,18 @@ class PreparedDataset(BaseDataset):
     y_scaler = None
     in_shape = None
     out_shape = None
-    class_set = None
-    class_counts = None
-    custom_splits = None
+    class_set = None # only filled if task='classification'
+    class_counts = None  # only filled if task='classification'
+    custom_splits = None # only filled if custom_splits are defined at data import
 
     def __init__(self, **kwargs) -> None:
+
         logger.info('Initializing PreparedClass for %s', kwargs['name'])
 
+        # initialize the BaseClass variables while keeping the_data in memory
         super().__init__(kwargs['meta_path'], kwargs['package_path'])
 
+        # storing relevant parameters
         self.in_components = kwargs['in_components']
         self.out_components = kwargs['out_components']
         self.in_chunk = kwargs['in_chunk']
@@ -98,116 +350,154 @@ class PreparedDataset(BaseDataset):
         self.slide_stride = kwargs['slide_stride']
         self.variable_sequence_length_limit = kwargs['variable_sequence_length_limit']
 
+        # Initiate the data preparation
         random.seed(self.seed)
         self._prepare()
 
     def get_dataset(self, set_tag: str, preload: bool = False):
+        """Creates and returns a PyTorch Dataset for a specified dataset split.
+
+        Parameters
+        ----------
+        set_tag : str
+            A string indicating the dataset split to load ('train', 'valid', 'eval').
+        preload : bool, default = False
+            Whether to preload the raw relevant instances and slice into memory when sampling feature values.
+
+        Returns
+        -------
+        Object
+            A PyTorch derived Dataset for the specified dataset split.
+        """
         if self.task == 'regression':
             if self.batch_sampling_mode in ('variable_length', 'variable_length_inference'):
-                logger.error(
-                    'Batch sampling mode %s is not supported for regression tasks. '
-                    'Did you mean to perform variable length regression (vl_regression)?',
-                    str(self.batch_sampling_mode)
-                )
+                logger.error('Batch sampling mode %s is not supported for regression tasks. '
+                             'Did you mean to perform variable length regression (vl_regression)?',
+                             str(self.batch_sampling_mode))
                 exit(101)
-            dataset = CustomDataset(
-                self.get_extractor(), self.selection[set_tag],
-                self.x_map, self.y_map,
-                self.x_scaler, self.y_scaler,
-                self.in_chunk, self.out_chunk,
-                self.padding_method, preload=preload
-            )
+            dataset = CustomDataset(self.get_extractor(), self.selection[set_tag],
+                          self.x_map, self.y_map,
+                          self.x_scaler, self.y_scaler,
+                          self.in_chunk, self.out_chunk,
+                          self.padding_method, preload=preload)
         elif self.task == 'classification':
             if self.batch_sampling_mode in ('variable_length', 'variable_length_inference'):
-                logger.error(
-                    'Batch sampling mode %s is not supported for classification tasks.',
-                    str(self.batch_sampling_mode)
-                )
+                logger.error('Batch sampling mode %s is not supported for classification tasks.',
+                             str(self.batch_sampling_mode))
                 exit(101)
-            dataset = CustomClassificationDataset(
-                self.get_extractor(), self.selection[set_tag],
-                self.x_map, self.y_map,
-                self.x_scaler, self.y_scaler,
-                self.in_chunk, self.out_chunk,
-                self.class_set,
-                self.padding_method,
-                preload=preload
-            )
+            dataset = CustomClassificationDataset(self.get_extractor(), self.selection[set_tag],
+                                                  self.x_map, self.y_map,
+                                                  self.x_scaler, self.y_scaler,
+                                                  self.in_chunk, self.out_chunk,
+                                                  self.class_set,
+                                                  self.padding_method,
+                                                  preload=preload)
         elif self.task == 'vl_regression':
-            dataset = CustomVariableLengthRegressionDataset(
-                self.get_extractor(), self.selection[set_tag],
-                self.x_map, self.y_map,
-                self.x_scaler, self.y_scaler,
-                self.in_chunk, self.out_chunk,
-                self.padding_method, preload=preload
-            )
+            dataset = CustomVariableLengthRegressionDataset(self.get_extractor(), self.selection[set_tag],
+                                    self.x_map, self.y_map,
+                                    self.x_scaler, self.y_scaler,
+                                    self.in_chunk, self.out_chunk,
+                                    self.padding_method, preload=preload)
         else:
             logger.error('Unknown task: %s', self.task)
             exit(101)
 
         return dataset
 
-    def get_dataloader(
-        self,
-        set_tag: str,
-        analysis: bool = False,
-        num_workers: int = 4,
-        preload: bool = False,
-        pin_memory: bool = True,
-        persistent_workers: bool | None = None,
-        prefetch_factor: int = 4,
-    ) -> DataLoader:
+
+    def get_dataloader(self, set_tag: str,
+                       analysis: bool = False,
+                       num_workers: int = 4,
+                       preload: bool = False) -> DataLoader:
+        """Creates and returns a PyTorch DataLoader for a specified dataset split.
+
+        This method generates a DataLoader for a given dataset split (e.g., train, valid, or eval).
+        It uses the `CustomDataset` class to create the dataset and then initializes
+        a DataLoader with a `CustomSampler` for batch generation.
+
+        Parameters
+        ----------
+        set_tag : str
+            A string indicating the dataset split to load ('train', 'valid', 'eval').
+        analysis : bool, default = False
+            A flag indicating whether the dataloader is being used for analysis purposes. If set to True,
+            the `drop_last` and `shuffle` parameters of the DataLoader will be set to False.
+        num_workers : int, default = 4
+            Sets the number of workers to use for loading the dataset.
+        preload : bool, default = False
+            Whether to preload the raw relevant instances and slice into memory when sampling feature values.
+
+        Returns
+        -------
+        DataLoader
+            A PyTorch DataLoader for the specified dataset split.
+
+        Notes
+        -----
+        If set_tag=`valid` or `eval` or analysis=True,
+        then the dataloader will not be shuffled, no smaller than batch_size batches will be dropped,
+        and batch_sampling_mode will be set to `inference`.
+        """
+
         if self.task == 'vl_regression' and self.batch_sampling_mode not in ('variable_length', 'variable_length_inference'):
-            logger.error(
-                'batch_sampling_mode=%s is not supported for variable length tasks. '
-                'Please ensure batch_sampling_mode=variable_length if task=vl_regression.',
-                str(self.batch_sampling_mode)
-            )
+            logger.error('batch_sampling_mode=%s is not supported for variable length tasks. '
+                         'Please ensure batch_sampling_mode=variable_length if task=vl_regression.',
+                         str(self.batch_sampling_mode))
             exit(101)
 
         if set_tag == 'train' and not analysis:
             shuffle = self.shuffle_train
             drop_small = True
-            sampler_mode = self.batch_sampling_mode
         else:
             shuffle = False
             drop_small = False
             if self.batch_sampling_mode == 'variable_length':
-                sampler_mode = 'variable_length_inference'
+                self.batch_sampling_mode = 'variable_length_inference'
             else:
-                sampler_mode = 'inference'
+                self.batch_sampling_mode = 'inference'
 
-        sampler = CustomSampler(
-            selection=self.selection[set_tag],
-            batch_size=self.batch_size,
-            input_size=self.in_shape[0],
-            seed=self.seed,
-            mode=sampler_mode,
-            drop_small=drop_small,
-            shuffle=shuffle,
-            slide_stride=self.slide_stride,
-            variable_sequence_length_limit=self.variable_sequence_length_limit
-        )
+        sampler = CustomSampler(selection=self.selection[set_tag],
+                                batch_size=self.batch_size,
+                                input_size=self.in_shape[0],
+                                seed=self.seed,
+                                mode=self.batch_sampling_mode,
+                                drop_small=drop_small,
+                                shuffle=shuffle,
+                                slide_stride=self.slide_stride,
+                                variable_sequence_length_limit=self.variable_sequence_length_limit)
 
         dataset = self.get_dataset(set_tag, preload=preload)
 
-        if persistent_workers is None:
-            persistent_workers = num_workers > 0
+        dataloader = DataLoader(dataset=dataset, batch_sampler=sampler, num_workers=num_workers)
 
-        dl_kwargs = {
-            "dataset": dataset,
-            "batch_sampler": sampler,
-            "num_workers": num_workers,
-            "pin_memory": pin_memory,
-            "persistent_workers": persistent_workers if num_workers > 0 else False,
-        }
-        if num_workers > 0:
-            dl_kwargs["prefetch_factor"] = prefetch_factor
-
-        dataloader = DataLoader(**dl_kwargs)
         return dataloader
 
+
     def get_ist_values(self, set_tag: str) -> list:
+        """Get the IST values for a given set tag.
+
+        This function retrieves the Instance-Slice-Time (IST) for each prediction point in a specific dataset split.
+
+        Parameters
+        ----------
+        set_tag : str
+            The data split to retrieve the IST values for.
+            Options are 'train', 'valid', 'eval' or 'all'.
+
+        Returns
+        -------
+        ist_values : list
+            The IST values for a given set tag.
+            Each row represents one prediction point by the following three values.
+                - The first value (Any) indicates the instance that the prediction point belongs to.
+                - The second value (int) indicates the index, within the instance, of the slice that the prediction point belongs to.
+                - The third value (numpy.datetime64) indicates the timestep at which the prediction point is found.
+
+        Notes
+        -----
+            - If set_tag='all' then all prediction points are returned in order ['train', 'valid', 'eval'].
+        """
+
         the_data = self.get_extractor()
         ist_values = []
         if set_tag == 'all':
@@ -217,7 +507,6 @@ class PreparedDataset(BaseDataset):
         else:
             logger.error('Unknown set_tag {}'.format(set_tag))
             exit(101)
-
         for tag in tags:
             for p in self.selection[tag]:
                 i = p[0]
@@ -229,6 +518,29 @@ class PreparedDataset(BaseDataset):
         return ist_values
 
     def fetch_input_points_manually(self, set_tag: str, point_ids: int | list) -> dict:
+        """ Manually fetch data points from the datamodule based on provided point IDs.
+
+        Parameters
+        ----------
+        set_tag : str
+            A string indicating the dataset split to load ('train', 'valid', 'eval').
+        point_ids : int | list[int]
+            The IDs of the data points to fetch. These indices are defined as the relative position in the selection
+            matrix corresponding to the set tag. Can be a single integer, a list of integers,
+            or a tuple specifying a range (start, end).
+
+        Returns
+        -------
+        dict
+            A dictionary containing the data points corresponding to the provided
+            IDs at key 'x'.
+
+        Raises
+        ------
+        ValueError
+            If the provided point IDs are invalid or out of range.
+
+        """
         dataset = self.get_dataset(set_tag, preload=False)
 
         try:
@@ -237,93 +549,107 @@ class PreparedDataset(BaseDataset):
                 custom_batch['x'] = unsqueeze(custom_batch['x'].contiguous(), 0)
                 custom_batch['y'] = unsqueeze(custom_batch['y'].contiguous(), 0)
         except ValueError:
-            logger.error(
-                'Invalid: ids %s not in choice "%s" (which has range %s)',
-                str(point_ids), set_tag, str((0, len(self.selection[set_tag])))
-            )
+            logger.error('Invalid: ids %s not in choice "%s" (which has range %s)',
+                         str(point_ids), set_tag,
+                         str((0, len(self.selection[set_tag]))))
             exit(101)
 
         return custom_batch
 
     def _prepare(self) -> None:
+        """Prepare the dataset by splitting and scaling the data.
+
+        This function calculates some relevant overhead variables (in_shape, out_shape, etc.),
+        it defines the correct data splits, and fits the correct scalers.
+
+        Notes
+        -----
+            - The splits are defined in the 'selection' variable in three selection matrices.
+            - The scalers are stored in the 'x_scaler' and 'y_scaler' variables.
+            - After the data is prepared the 'the_data' dictionary is removed from memory.
+
+        """
+        # check that desired components are in dataset
         missing_in_components = set(self.in_components) - set(self.components)
         if len(missing_in_components) > 0:
-            logger.error('Defined in_components %s not in data option.', str(missing_in_components))
+            logger.error('Defined in_components %s not in data option.',
+                         str(missing_in_components))
             exit(101)
-
         missing_out_components = set(self.out_components) - set(self.components)
         if len(missing_out_components) > 0:
-            logger.error('Defined out_components %s not in data option.', str(missing_out_components))
+            logger.error('Defined out_components %s not in data option.',
+                         str(missing_out_components))
             exit(101)
 
-        self.y_map = array([i for i, e in enumerate(self.components) if e in self.out_components])
-        self.x_map = array([i for i, e in enumerate(self.components) if e in self.in_components])
+        # infer input output component mappings
+        self.y_map = array([i for i, e in enumerate(self.components)
+                            if e in self.out_components])
+        self.x_map = array([i for i, e in enumerate(self.components)
+                            if e in self.in_components])
 
+        # infer model input and output shapes
         self.in_shape = [self.in_chunk[1] - self.in_chunk[0] + 1, len(self.in_components)]
         self.out_shape = [self.out_chunk[1] - self.out_chunk[0] + 1, len(self.out_components)]
-
         if self.task == 'classification':
             if self.out_chunk[0] != self.out_chunk[1]:
-                logger.error(
-                    'Currently, KnowIt can only perform classification at one specific time step at a time. '
-                    'Please change the out_chunk %s argument to reflect this. Both values must match.',
-                    str(self.out_chunk)
-                )
+                logger.error('Currently, KnowIt can only perform classification at one specific time step at a time. '
+                             'Please change the out_chunk %s argument to reflect this. Both values must match.',
+                             str(self.out_chunk))
                 exit(101)
             self._get_classes()
             self._count_classes()
             self.out_shape = [1, len(self.class_set)]
 
+        # split the dataset
         logger.info('Preparing data splits (selection).')
-        self.selection = DataSplitter(
-            self.get_extractor(),
-            self.split_method,
-            self.split_portions,
-            self.limit, self.y_map,
-            self.in_chunk, self.out_chunk,
-            self.min_slice,
-            custom_splits=self.custom_splits
-        ).get_selection()
+        self.selection = DataSplitter(self.get_extractor(),
+                                      self.split_method,
+                                      self.split_portions,
+                                      self.limit, self.y_map,
+                                      self.in_chunk, self.out_chunk,
+                                      self.min_slice,
+                                      custom_splits=self.custom_splits).get_selection()
 
         self.train_set_size = len(self.selection['train'])
         self.valid_set_size = len(self.selection['valid'])
         self.eval_set_size = len(self.selection['eval'])
-
-        logger.info('Data split sizes: %s', str((self.train_set_size, self.valid_set_size, self.eval_set_size)))
+        logger.info('Data split sizes: ' + str((self.train_set_size,
+                                                self.valid_set_size,
+                                                self.eval_set_size)))
 
         min_split = min(self.train_set_size, self.valid_set_size, self.eval_set_size)
         if self.batch_size > min_split:
-            logger.warning(
-                'Selected batch_size %s, is larger than one of dataset splits. Setting batch_size to %s.',
-                str(self.batch_size), str(min_split)
-            )
+            logger.warning('Selected batch_size %s, is larger than one of dataset splits. Setting batch_size to %s.',
+                           str(self.batch_size), str(min_split))
             self.batch_size = min_split
 
+        # scale the dataset
         logger.info('Preparing data scalers, if relevant.')
 
         if self.task == 'classification' and self.scaling_tag == 'full':
-            logger.warning(
-                'scaling_tag cannot be full for classification tasks. Changing to scaling_tag=in_only.'
-            )
+            logger.warning('scaling_tag cannot be full for classification tasks. Changing to scaling_tag=in_only.')
             self.scaling_tag = 'in_only'
 
-        self.x_scaler, self.y_scaler = DataScaler(
-            self.get_extractor(),
-            self.selection['train'],
-            self.scaling_method,
-            self.scaling_tag,
-            self.x_map,
-            self.y_map
-        ).get_scalers()
+        self.x_scaler, self.y_scaler = DataScaler(self.get_extractor(),
+                                                  self.selection['train'],
+                                                  self.scaling_method,
+                                                  self.scaling_tag,
+                                                  self.x_map,
+                                                  self.y_map).get_scalers()
 
     def _get_classes(self) -> None:
+        """Identify unique classes in the dataset.
+
+        This method processes the dataset to determine the unique classes present in the data.
+        The unique classes are stored in the `class_set` attribute.
+        """
+
         if self.task != 'classification':
             logger.error('Task must be classification to determine classes.')
             exit(101)
 
         data_extractor = self.get_extractor()
         found_class_set = set()
-
         for i in data_extractor.data_structure:
             vals = data_extractor.instance(i)
             subset = vals.iloc[:, self.y_map]
@@ -336,7 +662,8 @@ class PreparedDataset(BaseDataset):
                     unique_entries_list.append(tuple(u))
                 else:
                     unique_entries_list.append(u.item())
-            found_class_set.update(set(unique_entries_list))
+            unique_entries = unique_entries_list
+            found_class_set.update(set(unique_entries))
 
         self.class_set = {}
         tick = 0
@@ -344,10 +671,13 @@ class PreparedDataset(BaseDataset):
             self.class_set[c] = tick
             tick += 1
 
-        logger.info('Found %s unique classes.', str(len(self.class_set)))
+        logger.info('Found %s unique classes.',
+                    str(len(self.class_set)))
         logger.info(self.class_set)
 
     def _count_classes(self) -> None:
+        """ Count the number of instances of each class and store in a dictionary as attribute."""
+
         if self.task != 'classification' or not hasattr(self, 'class_set'):
             logger.error('Task must be classification and classes must have been determined to count classes.')
             exit(101)
@@ -369,19 +699,81 @@ class PreparedDataset(BaseDataset):
             self.class_counts[self.class_set[c]] = class_count
 
 
+
 class CustomSampler(Sampler):
-    def __init__(
-        self,
-        selection: array,
-        batch_size: int,
-        input_size: int,
-        seed: int = None,
-        mode: str = 'independent',
-        drop_small: bool = True,
-        shuffle: bool = True,
-        slide_stride: int = 1,
-        variable_sequence_length_limit: int = None
-    ) -> None:
+    """
+    Custom batch sampler for deep time series models, supporting different modes of temporal contiguity.
+
+    Parameters
+    ----------
+    selection : array, shape=[n_samples, 3]
+        Selection matrix containing instance IDs, slice IDs, time steps.
+    batch_size : int
+        Number of sequences per batch.
+    seed : int, default=None
+        Random seed for reproducibility.
+    mode : str, default='independent'
+        Either 'independent', 'sliding-window', 'inference', 'variable_length' or 'variable_length_inference' as described below.
+    drop_small : bool, default=True
+        Whether to drop batches smaller than batch_size.
+    shuffle : bool, default=True
+        Whether to apply shuffling.
+    slide_stride : int, default=1
+        The stride used for the sliding-window approach, if selected.
+    variable_sequence_length_limit : int, default=None
+        If non-None value given and 'mode' is either 'variable_length' or 'variable_length_inference',
+        the sequence lengths within a batch will be limited to the given value.
+
+    Attributes
+    ----------
+    selection : array, shape=[n_samples, 3]
+        Selection matrix containing instance IDs, slice IDs, time steps.
+    batch_size : int
+        Number of sequences per batch.
+    input_size : int
+        Number of time delays in the input of the model.
+    seed : int, default=None
+        Random seed for reproducibility.
+    mode : str, default='independent'
+        Either 'independent', 'sliding-window', 'inference', 'variable_length' or 'variable_length_inference' as described below.
+    drop_small : bool, default=True
+        Whether to drop batches smaller than batch_size.
+    shuffle : bool, default=True
+        Whether to apply shuffling.
+    slide_stride : int, default=1
+        The stride used for the sliding-window approach, if selected.
+    variable_sequence_length_limit : int, default=None
+        If non-None value given and 'mode' is either 'variable_length' or 'variable_length_inference',
+        the sequence lengths within a batch will be limited to the given value.
+    batches : list
+        The current set of batches that will be iterated over.
+    epoch : int
+        Current epoch. Used for random seeding.
+
+    Notes
+    -----
+        - mode='independent': Time is contiguous within sequences but not across batches.
+        - mode='sliding-window': A sliding window approach is used to ensure that time is contiguous within sequences and across batches.
+        - mode='inference': Same as 'sliding-window', but no shuffling, expansion for batch sizing, and striding.
+        - mode='variable_length': Similar to 'sliding-window' but the sequence lengths will be variable within batches.
+        - mode='variable_length_inference': Same as 'variable_length', but no shuffling or expansion for batch sizing.
+        - shuffle=False: Batches are constructed in dataset order as per the "selection" array. For variable length modes, slices will also be samples in descending order of length.
+        - shuffle=True:
+            - mode='independent': Sequences within and across batches are randomly shuffled.
+            - mode='sliding-window' or 'variable_length': Slices are shuffled before and after expansion, and a random number of prediction points (between 0 and 10) at the start of each slice are dropped before batches are constructed.
+    """
+
+    def __init__(self,
+                 selection: array,
+                 batch_size: int,
+                 input_size: int,
+                 seed: int = None,
+                 mode: str = 'independent',
+                 drop_small: bool = True,
+                 shuffle: bool = True,
+                 slide_stride: int = 1,
+                 variable_sequence_length_limit: int = None) -> None:
+
         self.selection = selection
         self.batch_size = batch_size
         self.input_size = input_size
@@ -397,16 +789,60 @@ class CustomSampler(Sampler):
         self.set_epoch(0)
 
     def __iter__(self):
+        """
+        Returns an iterator over the batches, and generates them if they do not exist.
+
+        Returns
+        -------
+        iterator
+            An iterator over the generated batches.
+        """
         if self.batches == []:
             self.set_epoch(self.epoch)
         return iter(self.batches)
 
     def __len__(self):
+        """
+        Returns the number of batches, and generates them if they do not exist.
+
+        Returns
+        -------
+        int
+            Number of generated batches.
+        """
         if self.batches == []:
             self.set_epoch(self.epoch)
         return len(self.batches)
 
     def set_epoch(self, epoch: int):
+        """
+        If the next epoch not already set, sets the next epoch and regenerate batches according to the sampling mode.
+
+        This method updates the internal epoch counter and triggers batch generation
+        based on the configured mode. It is intended to be called by the training loop
+        (e.g., a PyTorch Lightning model wrapper) at the end of each epoch.
+
+        Parameters
+        ----------
+        epoch : int
+            The next epoch number.
+
+        Raises
+        ------
+        SystemExit
+            If an unknown sampling mode is encountered.
+
+        Notes
+        -----
+        Supported modes:
+        - 'independent': Generates batches without enforcing temporal continuity.
+        - 'sliding-window': Generates batches using a sliding window approach for temporal consistency.
+        - 'inference': Prepares batches for model inference.
+        - 'variable_length': Prepares variable length batches for model training.
+        - 'variable_length_inference': Prepares variable length batches for model inference.
+
+        Additional checks are performed to ensure that batches meet size requirements.
+        """
         if epoch != self.epoch:
             self.epoch = epoch
             if self.mode == 'independent':
@@ -420,16 +856,24 @@ class CustomSampler(Sampler):
             elif self.mode == 'variable_length_inference':
                 self._create_vl_inference_batches()
             else:
-                logger.error(
-                    'Unknown sampler mode %s. Expected (independent, sliding-window, inference, '
-                    'variable_length, or variable_length_inference).',
-                    self.mode
-                )
+                logger.error('Unknown sampler mode %s. Expected (independent, sliding-window, inference, variable_length, or variable_length_inference).',
+                             self.mode)
                 exit(101)
+
+            # only for debugging
+            # self._batch_analyser()
 
             self._check_small()
 
     def _create_default_batches(self) -> None:
+        """
+        Creates batches in a simple sequential order. Shuffles if enabled.
+
+        Notes
+        -----
+        - The resulting batches adhere to the desired batch size up to the last edge-case, which will be smaller.
+        - The resulting batches will contain all appropriate prediction points exactly once.
+        """
         self.batches = []
         sampling = arange(len(self.selection))
         if self.shuffle:
@@ -440,72 +884,207 @@ class CustomSampler(Sampler):
             self.batches.append(batch)
 
     def _create_inference_batches(self) -> None:
+        """
+        Creates batches while maintaining temporal contiguity across batches, using a sliding window approach.
+        No shuffling, striding, or expansion (to fill batch size) is performed.
+
+        Notes
+        -----
+        - The resulting batches do not necessarily adhere to the desired batch size.
+        - The resulting batches will contain all appropriate prediction points exactly once.
+        """
         self.batches = []
         contiguous_slices = self._get_contiguous_slices()
 
+        # subsample slices if stride is more than one
         if self.slide_stride > 1:
-            logger.warning(
-                'Note: with sliding window stride greater than 1, inference is not done on all available prediction points.'
-            )
+            logger.warning('Note: with sliding window stride greater than 1, '
+                           'inference is not done on all available prediction points.')
             contiguous_slices = [s[::self.slide_stride] for s in contiguous_slices]
 
         self._block_sample_contiguous_batches(contiguous_slices)
 
     def _create_sliding_window_batches(self) -> None:
+        """
+        Creates batches while maintaining temporal contiguity across batches,
+        using a sliding window approach. Also shuffles, expands the batches to match the desired batch size,
+        and applies stride to sliding window if selected.
+        """
         self.batches = []
 
+        # if shuffle is on prepare rng
         rng = None
         if self.shuffle:
             rng = self._get_rng()
 
+        # retrieve contiguous slices
         contiguous_slices = self._get_contiguous_slices()
 
+        # shuffle slices if shuffle is on
         if self.shuffle:
             rng.shuffle(contiguous_slices)
 
+        # expand contiguous slices to satisfy batch size
         contiguous_slices = self._expand_contiguous_slices(contiguous_slices)
 
+        # shuffle expanded slices too, if shuffle is on
         if self.shuffle:
             rng.shuffle(contiguous_slices)
 
+        # if shuffle is on, drop a random number of prediction points at the start of slices
         if self.shuffle:
             contiguous_slices = self._random_drop_start(contiguous_slices, rng)
 
+        # subsample slices if stride is more than one
         if self.slide_stride > 1:
             contiguous_slices = [s[::self.slide_stride] for s in contiguous_slices]
 
+        # initiate batch sampling from contiguous slices
         self._block_sample_contiguous_batches(contiguous_slices)
 
     def _create_vl_batches(self) -> None:
+        """
+        Create and compile variable-length batches from contiguous slices.
+
+        This method prepares training batches by first retrieving
+        contiguous slices, expanding them to match the batch size, and
+        optionally applying several randomization steps if `self.shuffle` is enabled.
+        The resulting slices are then compiled into uniform-length mini-batches using
+        :meth:`_compile_vl_batches`.
+
+        The workflow is as follows:
+          1. Retrieve contiguous slices.
+          2. If shuffling is enabled:
+             - Initialize a random number generator (RNG).
+             - Shuffle the contiguous slices.
+          3. Expand slices to satisfy the batch size requirement.
+          4. If shuffling is enabled:
+             - Shuffle the expanded slices again.
+             - Apply a random offset to the start of slices (drop points).
+          5. Compile the prepared slices into batches.
+
+        Parameters
+        ----------
+        None
+
+        Returns
+        -------
+        None
+            The method updates the instance attribute `self.batches` in place.
+
+        Notes
+        -----
+        - Shuffling affects three stages: initial slices, expanded slices, and
+          optional random start-point dropping.
+        - The final batch construction is delegated to
+          :meth:`_compile_vl_batches`.
+
+        """
+
         self.batches = []
 
+        # if shuffle is on prepare rng
         rng = None
         if self.shuffle:
             rng = self._get_rng()
 
+        # retrieve contiguous slices
         contiguous_slices = self._get_contiguous_slices()
 
+        # shuffle slices if shuffle is on
         if self.shuffle:
             rng.shuffle(contiguous_slices)
 
+        # expand contiguous slices to satisfy batch size
         contiguous_slices = self._expand_contiguous_slices(contiguous_slices)
 
+        # shuffle expanded slices too, if shuffle is on
         if self.shuffle:
             rng.shuffle(contiguous_slices)
 
+        # if shuffle is on, drop a random number of prediction points at the start of slices
         if self.shuffle:
             contiguous_slices = self._random_drop_start(contiguous_slices, rng)
 
         self.batches = self._compile_vl_batches(contiguous_slices)
 
     def _create_vl_inference_batches(self) -> None:
+        """
+        Create variable-length batches for inference.
+
+        This method prepares batches specifically for inference by retrieving
+        contiguous slices, sorting them in descending order of
+        length, and compiling them into uniform-length mini-batches using
+        :meth:`_compile_vl_batches`.
+
+        Unlike training batch creation, no randomization or shuffling is applied.
+        Sorting by length ensures that longer sequences are grouped first, which
+        can improve efficiency during inference.
+
+        Parameters
+        ----------
+        None
+
+        Returns
+        -------
+        None
+            The method updates the instance attribute `self.batches` in place.
+
+        Notes
+        -----
+        - Sorting slices by length helps minimize truncation overhead
+          when compiling batches.
+        - The final batch construction is handled by :meth:`_compile_vl_batches`.
+
+        """
+
         self.batches = []
         contiguous_slices = self._get_contiguous_slices()
         contiguous_slices.sort(key=len, reverse=True)
         self.batches = self._compile_vl_batches(contiguous_slices)
 
     def _compile_vl_batches(self, slices: list) -> list:
+        """
+        Compile variable-length slices into uniform mini-batches.
+
+        This method groups a list of variable-length slices into batches of
+        equalized length by trimming them to the minimum length within each
+        candidate batch, or `self.variable_sequence_length_limit` if defined.
+        Remaining slice segments beyond the trimmed length
+        are recursively reprocessed until all input slices are consumed.
+
+        The batching process ensures that:
+          - Each batch contains slices truncated to the minimum length of
+            the batch (or `self.variable_sequence_length_limit`).
+          - Residual segments (after truncation) are re-queued for later batching.
+          - Empty slots in a candidate batch are filled with pending slices when possible.
+
+        Parameters
+        ----------
+        slices : list of list
+            A list of variable-length sequences (e.g., tokenized text, arrays, or
+            subsequences) to be grouped into uniform-length batches.
+
+        Returns
+        -------
+        list of list
+            A list of batches, where each batch is a list of truncated sequences
+            of equal length. Residual fragments are iteratively compiled until
+            no slices remain.
+
+        Notes
+        -----
+        - The batch size is determined by `self.batch_size`.
+        - Candidate batches may be smaller than `self.batch_size` if not enough
+          slices remain at the end of processing.
+
+        """
+
         def _fill_empty(batch, to_compile):
+            """
+            Replace empty sequences in a batch with items from `to_compile`.
+            Drop any that remain empty if no replacements are available.
+            """
             new_batch = []
             for seq in batch:
                 if len(seq) == 0:
@@ -519,7 +1098,8 @@ class CustomSampler(Sampler):
         to_compile = slices.copy()
 
         while to_compile:
-            candidate_batch, to_compile = _fill_empty(to_compile[:self.batch_size], to_compile[self.batch_size:])
+            candidate_batch, to_compile = _fill_empty(to_compile[:self.batch_size],
+                                                      to_compile[self.batch_size:])
             if candidate_batch:
                 if self.variable_sequence_length_limit is not None:
                     min_len = min(min(len(s), self.variable_sequence_length_limit) for s in candidate_batch)
@@ -533,13 +1113,119 @@ class CustomSampler(Sampler):
         return batches
 
     def _block_sample_contiguous_batches(self, contiguous_slices: list) -> None:
+        """
+        Generate and store batches from a list of contiguous slices.
+
+        This method processes a list of 1D NumPy arrays (contiguous slices) to create batches
+        of size `self.batch_size`. It iteratively samples blocks of slices, combines them with
+        remainders from previous iterations, handles empty slices, and constructs batches by
+        sampling across slices at the same position. The resulting batches are appended to
+        `self.batches`.
+
+        Parameters
+        ----------
+        contiguous_slices : list
+            A list of 1D NumPy arrays, each representing a contiguous block of prediction
+            points. Assumes at least `self.batch_size` slices are provided, though slices
+            may have variable lengths.
+
+        Returns
+        -------
+        None
+            The method appends generated batches to `self.batches` and does not return a value.
+
+        Raises
+        ------
+        ValueError
+            If `contiguous_slices` is empty or has fewer slices than `self.batch_size`.
+
+        Notes
+        -----
+        - The method assumes `contiguous_slices` contains at least `self.batch_size` slices.
+          If this condition is not met, it should be validated externally or an error raised.
+        - Slices may have variable lengths, and empty slices are replaced with the last
+          non-empty slice to maintain batch consistency, which may affect statefulness across
+          batches.
+        - Remainders from one iteration are combined with the next block of slices, which may
+          result in batches larger than `self.batch_size` if the combined block exceeds this
+          size.
+        - The process continues until all slices are processed or no non-empty slices remain.
+        - The implementation uses helper functions to sample blocks, concatenate remainders,
+          handle empty slices, and compute minimum lengths for clarity and modularity.
+        """
+
         def _sample_new_candidate_block(slices: list, idx: int) -> list:
+            """
+            Sample a block of slices from the given index to index + batch size.
+
+            Parameters
+            ----------
+            slices : list
+                List of slices to sample from.
+            idx : int
+                Starting index for sampling.
+
+            Returns
+            -------
+            list
+                A sublist of slices from `idx` to `idx + self.batch_size`, or an empty list
+                if the range is outside the bounds of `slices`.
+            """
             return slices[idx:idx + self.batch_size]
 
         def _concat_slices(a: list, b: list, c: int) -> list:
+            """
+            Concatenate corresponding elements from two lists of arrays up to a specified length.
+
+            Parameters
+            ----------
+            a : list
+                First list of arrays to concatenate.
+            b : list
+                Second list of arrays to concatenate.
+            c : int
+                Number of elements to concatenate from each list.
+
+            Returns
+            -------
+            list
+                A list of concatenated arrays, where each element is the concatenation of
+                corresponding arrays from `a` and `b` up to index `c`.
+            """
             return [concatenate((a[s], b[s])) for s in range(c)]
 
         def _combine_with_remainder(block: list, rem_block: list) -> list:
+            """
+            Combine a candidate block with a remainder block from a previous iteration.
+
+            Combines a new block of slices with a remainder block, handling cases where the
+            blocks have equal or different lengths. If the remainder exists, it concatenates
+            corresponding slices up to the shorter length and extends with excess slices from
+            the longer block.
+
+            Parameters
+            ----------
+            block : list
+                A list of arrays representing the new candidate block of slices.
+            rem_block : list or None
+                A list of arrays representing the remainder from the previous iteration, or None
+                if no remainder exists.
+
+            Returns
+            -------
+            list
+                A list of arrays combining the input block and remainder, or the input block
+                unchanged if no remainder is provided.
+
+            Notes
+            -----
+            - If `block` is empty and `rem_block` exists, returns `rem_block`.
+            - If `block` and `rem_block` have equal lengths, concatenates corresponding slices.
+            - If lengths differ, concatenates up to the shorter length and extends with excess
+              slices from the longer block (either `block` or `rem_block`).
+            - May result in larger-than-expected batches if the combined length exceeds the
+              intended batch size.
+            """
             if rem_block is not None:
                 if len(block) == 0:
                     block = rem_block
@@ -548,22 +1234,58 @@ class CustomSampler(Sampler):
                 elif len(block) != len(rem_block):
                     short = min(len(block), len(rem_block))
                     long = max(len(block), len(rem_block))
-                    combined = _concat_slices(rem_block, block, short)
+                    block = _concat_slices(rem_block, block, short)
                     if len(block) < len(rem_block):
-                        combined.extend([rem_block[b] for b in range(short, long)])
+                        block.extend([rem_block[b] for b in range(short, long)])
                     else:
-                        combined.extend([block[b] for b in range(short, long)])
-                    block = combined
+                        block.extend([block[b] for b in range(short, long)])
             return block
 
         def _get_min_len(block: list) -> int:
+            """
+            Compute the minimum length of arrays in a block of slices.
+
+            Parameters
+            ----------
+            block : list
+                A list of arrays representing a block of slices.
+
+            Returns
+            -------
+            int
+                The minimum length of any array in the block. Returns 0 if the block is empty.
+            """
             return min([len(s) for s in block])
 
         def _handle_zero_slices(block: list) -> list:
+            """
+            Replace empty slices in a block with the last non-empty slice.
+
+            If any slice in the block has zero length, it is replaced with the last non-empty
+            slice in the block. If the last slice is empty, it is removed. This process continues
+            until no empty slices remain or the block is empty.
+
+            Parameters
+            ----------
+            block : list
+                A list of arrays representing a block of slices.
+
+            Returns
+            -------
+            list
+                The modified block with empty slices replaced by the last non-empty slice or
+                an empty list if all slices are removed.
+
+            Notes
+            -----
+            - Replacing empty slices with the last non-empty slice may break statefulness
+              across batches for the affected indices.
+            - The process biases the last index toward lower contiguousness over time.
+            """
             if _get_min_len(block) == 0:
                 missing_entries = [_ for _ in range(len(block)) if len(block[_]) == 0]
                 while len(missing_entries) > 0:
-                    if missing_entries[-1] == len(block) - 1:
+                    if missing_entries[-1] == len(block)-1:
                         block.pop()
                         missing_entries.pop()
                     else:
@@ -591,16 +1313,81 @@ class CustomSampler(Sampler):
             next_block_at += self.batch_size
 
     def _random_drop_start(self, contiguous_slices: list, rng: Generator, drop_max: int = 10) -> list:
+        """
+        Randomly drop elements from the start of each slice in a list of contiguous slices.
+
+        This method randomly trims a number of elements from the beginning of each slice,
+        with the number of elements to drop chosen uniformly from 0 to the minimum of the
+        slice length and `drop_max`. The modified slices are returned.
+
+        Parameters
+        ----------
+        contiguous_slices : list
+            A list of 1D NumPy arrays, each representing a contiguous block of prediction
+            points.
+        rng : numpy.random.Generator
+            A random number generator for sampling the number of elements to drop.
+        drop_max : int, optional
+            The maximum number of elements to drop from the start of each slice (default is 10).
+
+        Returns
+        -------
+        list
+            The modified list of slices with random portions dropped from the start of each slice.
+
+        Notes
+        -----
+        - If a slice is shorter than `drop_max`, the number of elements dropped is limited to
+          the slice's length.
+        - Empty slices remain unchanged, as the minimum of their length and `drop_max` is 0.
+        """
         for s in range(len(contiguous_slices)):
             to_drop = rng.integers(0, min(len(contiguous_slices[s]), drop_max))
             contiguous_slices[s] = contiguous_slices[s][to_drop:]
+
         return contiguous_slices
 
     def _expand_contiguous_slices(self, contiguous_slices: list, dilation: int = 1) -> list:
+        """
+        Expand the list of contiguous slices to meet the batch size using a sliding window.
+
+        If the number of contiguous slices is less than the batch size, this method generates
+        additional slices by applying a sliding window with a specified dilation to existing
+        slices until the batch size is met or an error occurs. It handles cases where the
+        input list is empty or insufficient slices can be generated.
+
+        Parameters
+        ----------
+        contiguous_slices : list
+            A list of arrays, where each array contains indices representing a contiguous
+            block of prediction points from the selection matrix.
+        dilation : int, default=1
+            The dilation within a sampling window. Note this is different to `self.slide_stride`,
+            which defines the gaps between sampling windows.
+
+        Returns
+        -------
+        list
+            A list of arrays containing the original and newly generated contiguous slices,
+            expanded to meet or exceed the batch size.
+
+        Raises
+        ------
+        SystemExit
+            If the input `contiguous_slices` is empty or new slices cannot be generated
+            due to insufficient data for the dilation, logs an error and exits with code 101.
+
+        Notes
+        -----
+        The method uses a sliding window approach with a dilation defined by `dilation` to
+        generate new slices from existing ones. It iterates through the input slices, creating
+        new slices by shifting the starting point by the dilation value. A `found` flag tracks
+        whether new slices are successfully created in each iteration to avoid infinite loops.
+        If the batch size cannot be met, an error is logged, and the program exits.
+        """
         if len(contiguous_slices) == 0:
             logger.error('No contiguous slices to expand.')
             exit(101)
-
         if len(contiguous_slices) < self.batch_size:
             tick = 0
             found = False
@@ -615,46 +1402,92 @@ class CustomSampler(Sampler):
                     contiguous_slices.append(new_slice)
                     found = True
                 tick += 1
-
             if len(contiguous_slices) < self.batch_size:
-                logger.error(
-                    'Cannot construct ordered batches with current dataset, batch size %s and sliding-window-dilation %d',
-                    self.batch_size, dilation
-                )
+                logger.error('Cannot construct ordered batches with current dataset, '
+                             'batch size %s and sliding-window-dilation %d',
+                             self.batch_size, dilation)
                 exit(101)
-
         return contiguous_slices
 
     def _get_contiguous_slices(self) -> list:
+        """
+        Generate a list of contiguous slices from the selection matrix.
+
+        This method processes the selection matrix to identify unique instance-slice pairs,
+        gathers relevant prediction points for each pair, sorts them by time steps, and
+        splits them into contiguous blocks based on discontinuities in the indices. The
+        resulting blocks contain dataset-wide position indices.
+
+        Returns
+        -------
+        list
+            A list of arrays, where each array contains indices representing a contiguous
+            block of prediction points from the selection matrix.
+
+        Notes
+        -----
+        The method performs the following steps:
+            1. Indexes the selection matrix with dataset-wide position indices.
+            2. Identifies unique instance-slice pairs in the selection matrix.
+            3. For each pair, extracts and sorts prediction points by time steps.
+            4. Splits the sorted points into contiguous blocks based on index discontinuities.
+            5. Returns a list of these contiguous blocks, with each block containing the
+               dataset-wide position indices.
+        """
+
+        # add indices to selection matrix corresponding to dataset-wide position
         inx = expand_dims(arange(len(self.selection)), 1)
         selection = concatenate((self.selection, inx), axis=1)
 
+        # determine the unique set of instance-slices in the selection matrix
         instance_slice_pairs = unique(selection[:, :2], axis=0)
 
+        # determine the available set of contiguous slices
         contiguous_slices = []
         for s in instance_slice_pairs:
+            # get subselection
             idx = where((selection[:, 0] == s[0]) & (selection[:, 1] == s[1]))[0]
             s_block = selection[idx]
+
+            # sort according to time steps
             s_block = s_block[s_block[:, 2].argsort()]
+
+            # find a list of contiguous sub-blocks in subselection
             breakpoints = where(diff(s_block[:, 2]) != 1)[0] + 1
             s_block = split(s_block, breakpoints)
+
+            # store the dataset-wide indices as new contiguous slices
             s_block = [t[:, 3] for t in s_block]
             contiguous_slices.extend(s_block)
 
         return contiguous_slices
 
     def _check_small(self) -> None:
+        """
+        Drops small batches if drop_small is enabled and logs the number of dropped batches.
+        """
         total_before = len(self.batches)
         if self.drop_small:
             self.batches = [b for b in self.batches if len(b) >= self.batch_size]
         total_after = len(self.batches)
         if total_before != total_after:
-            logger.info(
-                '%s/%s batches dropped from dataloader when dropping batches smaller than %s.',
-                total_before - total_after, total_before, self.batch_size
-            )
+            logger.info('%s/%s batches dropped from dataloader when dropping batches smaller than %s.',
+                           total_before - total_after, total_before, self.batch_size)
 
     def _get_rng(self) -> Generator:
+        """
+        Returns a random number generator (RNG) instance based on the shuffle setting.
+
+        If shuffling is enabled, the RNG seed is adjusted by incorporating the current epoch
+        to ensure different randomization patterns across epochs. Otherwise, the RNG is
+        initialized with a fixed seed for reproducibility.
+
+        Returns
+        -------
+        random.Generator
+            A NumPy random number generator instance.
+
+        """
         if self.shuffle:
             rng = random.default_rng(self.seed + self.epoch)
         else:
@@ -662,6 +1495,9 @@ class CustomSampler(Sampler):
         return rng
 
     def _batch_analyser(self) -> None:
+
+        # ONLY FOR DEBUGGING
+
         num_pps = len(self.selection)
         batch_sizes = [len(b) for b in self.batches]
         coverage = []
@@ -700,19 +1536,44 @@ class CustomSampler(Sampler):
 
 
 class CustomDataset(Dataset):
-    def __init__(
-        self,
-        data_extractor,
-        selection_matrix,
-        x_map,
-        y_map,
-        x_scaler,
-        y_scaler,
-        in_chunk,
-        out_chunk,
-        padding_method,
-        preload: bool = False
-    ) -> None:
+    """
+    A custom dataset for deep time series models, using KnowIts data extraction protocols.
+
+    Parameters
+    ----------
+    data_extractor : DataExtractor
+        Object responsible for extracting data from disk.
+    selection_matrix : array
+        Matrix defining the selection of instances, slices, and time steps.
+    x_map : array, shape=[n_in_components,]
+        An array that contains the indices of BaseDataset.components that correspond to input components.
+    y_map : array, shape=[n_out_components,]
+        An array that contains the indices of BaseDataset.components that correspond to output components.
+    x_scaler : object
+        The scaler fitted to the training set input features.
+    y_scaler : object
+        The scaler fitted to the training set output features.
+    in_chunk : list, shape=[2,]
+        A list of two integers [a, b] for which a <= b,
+        defining the time steps of in_components for each prediction point.
+    out_chunk : list, shape=[2,]
+        A list of two integers [a, b] for which a <= b,
+        defining the time steps of out_components for each prediction point.
+    padding_method : str
+        The method to pad model inputs with.
+        Options can be found at (https://numpy.org/doc/stable/reference/generated/numpy.pad.html).
+        See heading for description.
+    preload : bool, default=False
+        Whether to preload the dataset into memory.
+
+    """
+    def __init__(self, data_extractor, selection_matrix,
+                 x_map, y_map,
+                 x_scaler, y_scaler,
+                 in_chunk, out_chunk,
+                 padding_method,
+                 preload: bool = False) -> None:
+
         self.data_extractor = data_extractor
         self.selection_matrix = selection_matrix
         self.x_map = x_map
@@ -725,111 +1586,140 @@ class CustomDataset(Dataset):
         self.preload = preload
 
         if self.in_chunk[0] == self.in_chunk[1] and self.padding_method not in ('constant', 'empty'):
-            logger.error(
-                'Data padding method %s with an input chunk size of 1 not supported. '
-                'Choose padding method from (constant, empty).',
-                self.padding_method
-            )
+            logger.error('Data padding method %s with an input chunk size of 1 not supported. '
+                           'Choose padding method from (constant, empty).', self.padding_method)
             exit(101)
 
         self.preloaded_slices = {}
         if preload:
-            logger.info(
-                "Preloading relevant slices into memory as NumPy arrays. "
-                "This could take a while, but should speed up actual training."
-            )
+            logger.info("Preloading relevant slices into memory. This could take a while, but speed up actual training.")
             instances = unique(self.selection_matrix[:, 0], axis=0)
             for i in instances:
                 slices = unique(self.selection_matrix[self.selection_matrix[:, 0] == i, 1])
                 for s in slices:
-                    slice_vals = self.data_extractor.slice(i, s)
-                    x_full = slice_vals.iloc[:, self.x_map].to_numpy(copy=True)
-                    y_full = slice_vals.iloc[:, self.y_map].to_numpy(copy=True)
-                    self.preloaded_slices[(i, s)] = (x_full, y_full)
+                    self.preloaded_slices[(i, s)] = self.data_extractor.slice(i, s)
 
     def __len__(self) -> int:
+        """Return the number of samples in the dataset.
+
+        Returns:
+        --------
+        int
+            The number of samples in the dataset.
+        """
         return self.selection_matrix.shape[0]
 
-    @staticmethod
-    def _normalize_idx(idx: int | list | Tensor | ndarray) -> list:
+    def __getitem__(self, idx: int | list | Tensor) -> dict:
+        """Return a single sample from the dataset at the given index.
+
+        Parameters
+        ----------
+        idx : int | list | Tensor
+            The index of the sample to retrieve or list of indices.
+
+        Returns
+        -------
+        dict[str, any]
+            A dictionary containing
+                -   'x' (Tensor): the input features,
+                -   'y' (Tensor): the target features,
+                -   's_id' (int): the sample ID 's_id'; it's index in the selection matrix.
+                -   'ist_idx' (array): the sample IST index.
+        """
+
         if type(idx) is list:
-            return idx
-        if is_tensor(idx):
-            return idx.tolist()
-        if isinstance(idx, ndarray):
-            return idx.tolist()
-        return [idx]
+            idx_list = idx
+        elif is_tensor(idx):
+            idx_list = idx.tolist()
+        elif isinstance(idx, ndarray):
+            idx_list = idx.tolist()
+        else:
+            # assumes idx is an integer
+            idx_list = [idx]
 
-    def _get_slice_arrays(self, selection, batch_cache: dict | None = None):
-        key = (selection[0], selection[1])
-
-        if self.preload:
-            return self.preloaded_slices[key]
-
-        if batch_cache is not None and key in batch_cache:
-            return batch_cache[key]
-
-        slice_vals = self.data_extractor.slice(selection[0], selection[1])
-        x_full = slice_vals.iloc[:, self.x_map].to_numpy(copy=False)
-        y_full = slice_vals.iloc[:, self.y_map].to_numpy(copy=False)
-
-        if batch_cache is not None:
-            batch_cache[key] = (x_full, y_full)
-
-        return x_full, y_full
-
-    def _make_single_sample(self, pp: int, batch_cache: dict | None = None) -> dict:
-        selection = self.selection_matrix[pp]
-        x_full, y_full = self._get_slice_arrays(selection, batch_cache=batch_cache)
-
-        x_vals = self._sample_and_pad(x_full, selection, self.in_chunk, self.padding_method)
-        y_vals = y_full[selection[2] + self.out_chunk[0]: selection[2] + self.out_chunk[1] + 1, :]
-
-        x_vals = self.x_scaler.transform(x_vals)
-        y_vals = self.y_scaler.transform(y_vals)
-
-        return self._package_output(x_vals, y_vals, pp, [selection])
-
-    def __getitem__(self, idx: int | list | Tensor | ndarray) -> dict:
-        idx_list = self._normalize_idx(idx)
-
-        if len(idx_list) == 1:
-            return self._make_single_sample(idx_list[0], batch_cache=None)
-
-        # Manual multi-fetch path, preserving previous behaviour
-        batch_cache = {}
         input_x = []
         output_y = []
         ist_idx = []
-
         for pp in idx_list:
+            # get sample selection
             selection = self.selection_matrix[pp]
+
+            # save sample ist index
             ist_idx.append(selection)
 
-            x_full, y_full = self._get_slice_arrays(selection, batch_cache=batch_cache)
+            # get relevant slice information
+            if self.preload:
+                slice_vals = self.preloaded_slices[(selection[0], selection[1])]
+            else:
+                slice_vals = self.data_extractor.slice(selection[0], selection[1])
 
-            x_vals = self._sample_and_pad(x_full, selection, self.in_chunk, self.padding_method)
-            y_vals = y_full[selection[2] + self.out_chunk[0]: selection[2] + self.out_chunk[1] + 1, :]
+            # get input values and pad if necessary
+            x_vals = slice_vals.iloc[:, self.x_map].to_numpy()
+            x_vals = self._sample_and_pad(x_vals, selection, self.in_chunk, self.padding_method)
 
+            # get output values (no padding allowed)
+            y_vals = slice_vals.iloc[:, self.y_map].to_numpy()
+            y_vals = y_vals[selection[2] + self.out_chunk[0]: selection[2] + self.out_chunk[1] + 1, :]
+
+            # scale inputs and outputs if applicable
             input_x.append(self.x_scaler.transform(x_vals))
             output_y.append(self.y_scaler.transform(y_vals))
 
-        input_x = array(input_x)
-        output_y = array(output_y)
+        if len(idx_list) > 1:
+            input_x = array(input_x)
+            output_y = array(output_y)
+        else:
+            input_x = input_x[0]
+            output_y = output_y[0]
 
         return self._package_output(input_x, output_y, idx, ist_idx)
 
-    def __getitems__(self, idx: list[int]) -> list[dict]:
-        """
-        Fast-path used by torch DataLoader when auto-collating batches.
-        Must return a list of individual samples, not an already-collated batch.
-        """
-        idx_list = self._normalize_idx(idx)
-        batch_cache = {}
-        return [self._make_single_sample(pp, batch_cache=batch_cache) for pp in idx_list]
-
     @staticmethod
     def _sample_and_pad(slice_vals, selection, s_chunk, pad_mode):
+        """
+        Sample a block of time series components from a slice and pad as needed.
+
+        This method extracts a sub-sequence of time series data from `slice_vals`
+        according to the range specified in `s_chunk`, anchored at the slice index
+        indicated in `selection`. If the requested range extends beyond the available
+        time steps, the resulting array is padded according to `pad_mode`.
+
+        Parameters
+        ----------
+        slice_vals : ndarray of shape (num_time_steps, num_in_components)
+            The full time series data for the slice, where rows correspond to time
+            steps and columns to different input components.
+
+        selection : ndarray of shape (3,)
+            A selection descriptor, where the third entry (`selection[2]`) represents
+            the index of the current time step around which to sample.
+
+        s_chunk : ndarray of shape (2,) or array-like
+            A two-element array specifying the relative range of time steps to
+            sample, inclusive. For example, `[-2, 2]` samples a window of 5
+            consecutive steps centered at `selection[2]`.
+
+        pad_mode : str
+            Padding mode passed to `numpy.pad`. Common modes include `'constant'`,
+            `'edge'`, and `'reflect'`.
+
+        Returns
+        -------
+        vals : ndarray of shape (window_size, num_in_components)
+            The sampled sub-sequence of time series components, padded if the
+            requested window extends outside the bounds of `slice_vals`. The
+            window size is defined as `s_chunk[1] - s_chunk[0] + 1`.
+
+        Notes
+        -----
+        - If the requested range falls completely outside the available time
+          steps, the method returns an array of shape `(window_size, num_in_components)`
+          consisting entirely of padded values.
+        - Padding is applied only along the time dimension. No padding occurs
+          along the component dimension.
+        - This method assumes `pad` refers to `numpy.pad`.
+        """
+
         far_left = 0
         far_right = slice_vals.shape[0]
         left = selection[2] + s_chunk[0]
@@ -855,51 +1745,43 @@ class CustomDataset(Dataset):
 
     @staticmethod
     def _package_output(input_x, output_y, idx, ist_idx):
-        sample = {
-            'x': from_numpy(input_x).float(),
-            'y': from_numpy(output_y).float(),
-            's_id': idx,
-            'ist_idx': ist_idx
-        }
+        """ Package the sample values for a basic regression problem. """
+        sample = {'x': from_numpy(input_x).float(),
+                  'y': from_numpy(output_y).float(),
+                  's_id': idx, 'ist_idx': ist_idx}
         return sample
 
 
 class CustomClassificationDataset(CustomDataset):
+    """A custom dataset for deep time series classification models, using KnowIt's data extraction protocols.
+    Inherits from CustomDataset.
+    """
     class_set = {}
     class_counts = {}
 
-    def __init__(
-        self,
-        data_extractor,
-        selection_matrix,
-        x_map,
-        y_map,
-        x_scaler,
-        y_scaler,
-        in_chunk,
-        out_chunk,
-        class_set,
-        padding_method,
-        preload: bool = False
-    ) -> None:
+    def __init__(self, data_extractor, selection_matrix,
+                 x_map, y_map,
+                 x_scaler, y_scaler,
+                 in_chunk, out_chunk,
+                 class_set, padding_method,
+                 preload: bool = False) -> None:
+
         if out_chunk[0] != out_chunk[1]:
-            logger.error(
-                'Currently, KnowIt can only perform classification at one specific time step at a time. '
-                'Please change the out_chunk %s argument to reflect this. Both values must match.',
-                str(out_chunk)
-            )
+            logger.error('Currently, KnowIt can only perform classification at one specific time step at a time. '
+                         'Please change the out_chunk %s argument to reflect this. Both values must match.',
+                         str(out_chunk))
             exit(101)
 
-        super().__init__(
-            data_extractor, selection_matrix, x_map, y_map,
-            x_scaler, y_scaler,
-            in_chunk, out_chunk,
-            padding_method,
-            preload
-        )
+        super().__init__(data_extractor, selection_matrix, x_map, y_map,
+                         x_scaler, y_scaler,
+                         in_chunk, out_chunk,
+                         padding_method,
+                         preload)
         self.class_set = class_set
 
     def _package_output(self, input_x, output_y, idx, ist_idx):
+        """ Package the sample values for a basic classification problem. """
+
         if len(output_y.shape) == 2:
             new_output_y = zeros_tensor(len(self.class_set))
             new_output_y[self.class_set[output_y.item()]] = 1
@@ -909,29 +1791,25 @@ class CustomClassificationDataset(CustomDataset):
             new_output_y[:, output_y] = 1
             output_y = new_output_y
 
-        sample = {
-            'x': from_numpy(input_x).float(),
-            'y': output_y,
-            's_id': idx,
-            'ist_idx': ist_idx
-        }
+        sample = {'x': from_numpy(input_x).float(),
+                  'y': output_y,
+                  's_id': idx, 'ist_idx': ist_idx}
+
         return sample
 
 
 class CustomVariableLengthRegressionDataset(CustomDataset):
-    def __init__(
-        self,
-        data_extractor,
-        selection_matrix,
-        x_map,
-        y_map,
-        x_scaler,
-        y_scaler,
-        in_chunk,
-        out_chunk,
-        padding_method,
-        preload: bool = False
-    ) -> None:
+    """A custom dataset for deep time series regression models that take variable length input, using KnowIt's data extraction protocols.
+    Inherits from CustomDataset.
+    """
+
+    def __init__(self, data_extractor, selection_matrix,
+                 x_map, y_map,
+                 x_scaler, y_scaler,
+                 in_chunk, out_chunk,
+                 padding_method,
+                 preload: bool = False) -> None:
+
         if in_chunk[0] != in_chunk[1]:
             logger.error('For variable length input modeling, input chunk size must be 1.')
             exit(101)
@@ -940,23 +1818,22 @@ class CustomVariableLengthRegressionDataset(CustomDataset):
             logger.error('For variable length input modeling, output chunk size must be 1.')
             exit(101)
 
-        super().__init__(
-            data_extractor, selection_matrix, x_map, y_map,
-            x_scaler, y_scaler,
-            in_chunk, out_chunk,
-            padding_method,
-            preload
-        )
+        super().__init__(data_extractor, selection_matrix, x_map, y_map,
+                         x_scaler, y_scaler,
+                         in_chunk, out_chunk,
+                         padding_method,
+                         preload)
 
     def _package_output(self, input_x, output_y, idx, ist_idx):
+        """ Package the sample values for a basic variable length regression problem. """
+
         if len(input_x.shape) == 3:
             input_x = input_x.squeeze(axis=1)
             output_y = output_y.squeeze(axis=1)
 
-        sample = {
-            'x': from_numpy(input_x).float(),
-            'y': from_numpy(output_y).float(),
-            's_id': idx,
-            'ist_idx': ist_idx
-        }
+        sample = {'x': from_numpy(input_x).float(),
+                  'y': from_numpy(output_y).float(),
+                  's_id': idx, 'ist_idx': ist_idx}
+
         return sample
+
